+++ tests/Microsoft.Identity.Test.Common/TestConstants.cs
@@ -179,6 +179,7 @@ m1t9gRT1mNeeluL4cZa6WyVXqXc6U2wfR5DY6GOMUubN5Nr1n8Czew8TPfab4OG37BuEMNmBpqoRrRgF
         }
 
         public const string Bearer = "Bearer";
+        public const string Pop = "PoP";
 
         public static IDictionary<string, string> ExtraQueryParameters
         {
